# fix: Unloaded lazy vector in duplicate field project