The following error was noticed when running a query as below, where the filter 
expressions were for one field and duplicate projects were for another field.

> SELECT id, id FROM tmp WHERE name = 'John'

The lazy vector is not preloaded because it’s only referenced in identity 
projections, not in expressions, and the `fillOutput` calls `wrapInDictionary` 
multiple times on the same lazy vector. The lazy vector is still unloaded at 
the time, and thus, the dictionary vector wraps the same underlying lazy vector 
multiple times while unloaded.

This PR fixes this issue by caching the already wrapped children to avoid 
wrapping the same child multiple times.

```
E20250626 11:53:16.482720 23136090 Exceptions.h:66] Line: /Users/velox/./velox/vector/DictionaryVector-inl.h:38, Function:setInternalState, Expression: dictionaryValues_->markAsContainingLazyAndWrapped() An unloaded lazy vector cannot be wrapped by two different top level vectors., Source: RUNTIME, ErrorCode: INVALID_STATE
unknown file: Failure
C++ exception with description "Exception: VeloxRuntimeError
Error Source: RUNTIME
Error Code: INVALID_STATE
Reason: An unloaded lazy vector cannot be wrapped by two different top level vectors.
Retriable: False
Expression: dictionaryValues_->markAsContainingLazyAndWrapped()
Context: Operator: FilterProject[2] 1
Function: setInternalState
File: /Users/velox/./velox/vector/DictionaryVector-inl.h
Line: 38
Stack trace:
```

Benchmark results:

Before fix:

```
duplicate_project_integer                                   7.88ms    126.97
duplicate_project_string                                   57.75ms     17.32
filter_integers_duplicate_project_integers                 10.19ms     98.15
filter_strings_duplicate_project_strings                   47.57ms     21.02
```

After fix:

```
duplicate_project_integer                                   7.68ms    130.17
duplicate_project_string                                   55.27ms     18.09
filter_integers_duplicate_project_integers                 11.24ms     89.00
filter_strings_duplicate_project_strings                   47.55ms     21.03
```