@kKPulla 2025-07-31
@rui-mo I see a test failure in the signals - https://github.com/facebookincubator/velox/actions/runs/16654290065/job/47135141462?pr=13895 Can you please fix it and I will update the PR to merge. Thanks!