@rui-mo 2025-08-01
@kKPulla Thanks for the pointer. I fixed the test failure on debug build. Now the CI fails due to https://github.com/facebookincubator/velox/issues/14309.